requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
ultralytics>=8.0.0
requests>=2.25.1
